Create separated modal component to show warning message
