+++ src/components/Main/index.ts
@@ -1,7 +1,5 @@
 import AppMenu from './AppMenu'
 import Footer from './Footer'
+import ModalWarning from './ModalWarning'
 
-export {
-    AppMenu,
-    Footer
-}
+export { AppMenu, Footer, ModalWarning }
