+++ src/containers/Application/NavigationBox.tsx
@@ -11,8 +11,7 @@ interface NavigationBoxProps {
   serialNumber: string
   currentPage: number
   validateElementsInPage: (props?: CurrentPage) => boolean
-  showValidationModal: Function
-  modalState: { showModal: ModalProps; setShowModal: Function }
+  setShowModal: (props: ModalProps) => void
 }
 
 const NavigationBox: React.FC<NavigationBoxProps> = ({
@@ -21,7 +20,7 @@ const NavigationBox: React.FC<NavigationBoxProps> = ({
   serialNumber,
   currentPage,
   validateElementsInPage,
-  modalState: { setShowModal },
+  setShowModal,
 }) => {
   const nextSection = templateSections.find(({ index }) => index === currentSection.index + 1)
   const previousSection = templateSections.find(({ index }) => index === currentSection.index - 1)
