+++ src/utils/constants.ts
@@ -3,6 +3,7 @@ export default {
   BUTTON_BACK_DASHBOARD: 'Back to dashboard',
   BUTTON_BACK_TO: 'Back to',
   BUTTON_NEXT: 'Next',
+  BUTTON_OPTION_YES: 'Yes',
   BUTTON_PREVIOUS: 'Previous',
   BUTTON_REVIEW_APPROVE_ALL: 'Approve all ',
   BUTTON_REVIEW_CONTINUE: 'Continue review',
@@ -23,6 +24,7 @@ export default {
   ERROR_REVIEW_PAGE: 'Problem to load review',
   ERROR_REVIEW_OVERVIEW: 'Problem to load review homepage',
   ERROR_LOGIN_PASSWORD: 'Oops! Problem with username or password',
+  ERROR_SUBMISSION_INVALID: 'Problem to submit application',
   LABEL_ASSIGNED_TO_YOU: 'Assigned to you',
   LABEL_ASSIGNED_TO_OTHER: 'Assigned to another reviewer',
   LABEL_LOGIN_PASSWORD: 'Password',
